move tree_hash() into its own crate (clvm-utils)
