+++ .github/workflows/build-crate-and-npm.yml
@@ -43,7 +43,10 @@ jobs:
     - name: build
       run: cargo build --release
     - name: dry-run of `cargo publish`
-      run: cargo publish --dry-run
+      run: |
+        cargo publish --dry-run
+        cd clvm-utils
+        cargo publish --dry-run
 
     - name: Upload crate artifacts
       uses: actions/upload-artifact@v2
@@ -68,7 +71,10 @@ jobs:
       if: startsWith(github.event.ref, 'refs/tags')
       env:
         CARGO_REGISTRY_TOKEN: ${{ secrets.cargo_registry_token }}
-      run: cargo publish
+      run: |
+        cargo publish
+        cd clvm-utils
+        cargo publish
 
     # this has not been tested, so probably needs to be debugged next time a tag is created
     - name: publish to npmjs.com if tagged
