+++ Cargo.toml
@@ -1,7 +1,7 @@
 # the "wheel" crate is excluded from the workspace because pyo3 has problems with
 # "cargo test" and "cargo bench"
 [workspace]
-members = ["wasm", "chia_streamable_macro", "chia-bls"]
+members = ["wasm", "chia_streamable_macro", "chia-bls", "clvm-utils"]
 exclude = ["wheel"]
 
 [package]
@@ -22,6 +22,7 @@ clvmr = "=0.1.24"
 hex = "=0.4.3"
 pyo3 = { version = "=0.15.1", features = ["extension-module"], optional = true }
 chia_streamable_macro = { version = "=0.2.3", path = "chia_streamable_macro" }
+clvm-utils = { version = "=0.1.14", path = "clvm-utils" }
 
 [dev-dependencies]
 num-traits = "=0.2.15"
