+++ clvm-utils/README.md
@@ -1,0 +1,9 @@
+Library providing utility functions for handling CLVM programs
+
+
+computing tree hash:
+
+```rust
+fn tree_hash(a: &Allocator, node: NodePtr) -> [u8; 32]
+```
+
