+++ clvm-utils/fuzz/Cargo.toml
@@ -1,0 +1,29 @@
+[package]
+name = "clvm-utils-fuzz"
+version = "0.0.0"
+authors = ["Automatically generated"]
+publish = false
+edition = "2021"
+
+[package.metadata]
+cargo-fuzz = true
+
+[dependencies]
+libfuzzer-sys = "0.4"
+clvmr = "=0.1.24"
+
+[dependencies.clvm-utils]
+path = ".."
+
+[dependencies.chia]
+path = "../.."
+
+# Prevent this from interfering with workspaces
+[workspace]
+members = ["."]
+
+[[bin]]
+name = "tree-hash"
+path = "fuzz_targets/tree-hash.rs"
+test = false
+doc = false
