+++ clvm-utils/fuzz/fuzz_targets/tree-hash.rs
@@ -3,7 +3,7 @@ use libfuzzer_sys::fuzz_target;
 
 use clvmr::allocator::Allocator;
 use chia::fuzzing_utils::{BitCursor, make_tree};
-use chia::gen::get_puzzle_and_solution::tree_hash;
+use clvm_utils::tree_hash::tree_hash;
 
 fuzz_target!(|data: &[u8]| {
     let mut a = Allocator::new();
